将机器人作为单独模块分离
